requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
python-multipart>=0.0.6
httpx>=0.25.0
requests>=2.31.0
python-dotenv>=1.0.0
colorama>=0.4.6
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0
cairosvg>=2.7.0
cloudinary>=1.36.0
pdf2image>=1.16.0
pytesseract>=0.3.10
